+++ next-docs/public/examples/table/SelectableCheckboxes.tsx
@@ -72,7 +72,7 @@ const Example = () => {
         progress: <span>{Math.floor(index * 100)}</span>,
         status: Math.floor(index * 100),
         activity: Math.floor(index * 100),
-        isSelected: index === 3
+        isSelected: index === 3,
       };
     });
   };