+++ next-docs/types/definition/core.d.ts
@@ -1,2 +1,1 @@
 declare module '@heathmont/moon-core-tw';
-declare module '@heathmont/moon-table-tw';