+++ workspaces/core/src/bottomSheet/private/types/TitleProps.ts
@@ -1,0 +1,6 @@
+type TitleProps = {
+  className?: string;
+  children?: React.ReactNode;
+};
+
+export default TitleProps;