+++ workspaces/core/src/breadcrumb/private/types/BreadcrumbProps.ts
@@ -1,5 +1,5 @@
 type BreadcrumbProps = {
-  breadcrumbs: React.ElementType[];
+  breadcrumbs: React.ReactNode[];
 };
 
 export default BreadcrumbProps;