+++ packages/components/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## 9.19.0
+
+### Minor Changes
+
+- Bump minor version
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @heathmont/moon-assets@9.19.0
+  - @heathmont/moon-icons@9.19.0
+  - @heathmont/moon-themes@9.19.0
+  - @heathmont/moon-utils@9.19.0
+
 ## 9.18.1
 
 ### Patch Changes